Move layers and rename
